+++ examples/bun/typescript/pool.ts
@@ -10,6 +10,11 @@ const workerFileURL = new URL('./worker.ts', import.meta.url)
 const fixedPool = new FixedThreadPool<MyData, MyResponse>(
   availableParallelism(),
   workerFileURL,
+  {
+    errorEventHandler: (e: ErrorEvent) => {
+      console.error(e)
+    },
+  },
 )
 
 await fixedPool.execute()
@@ -18,6 +23,11 @@ const dynamicPool = new DynamicThreadPool<MyData, MyResponse>(
   Math.floor(availableParallelism() / 2),
   availableParallelism(),
   workerFileURL,
+  {
+    errorEventHandler: (e: ErrorEvent) => {
+      console.error(e)
+    },
+  },
 )
 
 await dynamicPool.execute()